factored out log methods from class
